nvim: added python3 venv support, splits, folding
Neovim's Python providers don't work well within virtual environments,
so one needs to specify a Python installation that has the `neovim` pip
package installed.

The default splitting options for Vim will put a new window to the left
and to the top, so to keep in line with how i3 tiles windows, I set the
new windows to appear at the right and at the bottom, respectively for
vertical and horizontal splits.

Code folding is a nice feature that most graphical IDEs have, in order
to hide functions that don't need to be edited out of view. I haven't
set it up in neovim yet, surprisingly.